Updated the controller actions for the mobi
The goddamn mobile template renderings were a pain in the arse to find.
I’ve finally worked it and have control over the template details. I’m
now in a position to render a fully working version of the current site
(sans the ask question button, hopefully this will be okay tho…)