Updated the footer menu code / Added Store
This was a bugger to track down, and doesn’t quite work the way that I
intended it to work. I’ll have to chase it down a little more to work
out why. I’ve also updated the ErrorController because I ran into a
similar page not found error which was related to a missing controller.
Also found a store module for social engine and installed this, still
needs configuration, but it’s a pretty great start for the store